fix type error

+++ src/devicesController/desktopController.go
@@ -21,7 +21,7 @@ var workDesktopBrightnessCtrlPQ = utils.NewPriorityQueue()
 // Pings desktop status every <frequency> secoends
 // Updates brightness control pq
 // Sends brightness control command when changing offline to online
-func UpdateSystemStatus(wg *sync.WaitGroup, brightness int, systemAddress string, frequency int) {
+func UpdateSystemStatus(wg *sync.WaitGroup, brightness int, systemAddress string, frequency float32) {
 	pqElementName := "offline"
 	for {
 		_, err := utils.SearchPQElement(workDesktopBrightnessCtrlPQ, pqElementName)
